condaenv.v_b3s0p2.requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
apache-libcloud==3.8.0
argcomplete==3.6.2
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.40.2
botocore==1.40.2
brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
circuitbreaker==2.1.3
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
composer==0.31.0
contourpy==1.3.2
coolname==2.2.0
cramjam==2.11.0
cryptography==44.0.3
cycler==0.12.1
databricks-sdk==0.47.0
datasets==3.5.1
dill==0.3.8
docker==7.1.0
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.13.1
flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2023.6.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.15.1
llm-foundry==0.20.0
mako==1.3.10
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mlflow==2.21.3
mlflow-skinny==2.21.3
mosaicml==0.30.0
mosaicml-cli==0.7.3
mosaicml-streaming==0.12.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
oci==2.157.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==24.2
pandas==2.3.1
paramiko==3.5.1
peft==0.17.0
pillow==11.0.0
pip==25.2
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.5.0
pynvml==11.5.3
pyopenssl==24.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytorch-ranger==0.1.1
pytz==2025.2
pyyaml==6.0.2
questionary==2.1.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rsa==4.9.1
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.34.1
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slack-sdk==3.36.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
sqlparse==0.5.3
starlette==0.47.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch-optimizer==0.3.0
torchmetrics==1.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
wandb==0.18.7
wcwidth==0.2.13
websockets==11.0.3
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstd==1.5.7.2